move pytests directory to tests directory
